refactor/rename mag-ja.ts to mag-com.ts
