+++ src/app/metadata.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import { msg } from "@/msg-ja"
+import { msg } from "@/msg-com"
 export const metadata: Metadata = {
   title: msg.APP_NAME,
   description: "Numazu Mathematics Assignment Submission and Communication Hub",
