+++ src/compornents/ClassList.tsx
@@ -25,7 +25,7 @@ import FolderIcon from '@mui/icons-material/Folder';
 import { User } from "@prisma/client";
 import { useRouter } from "next/navigation";
 import { TeacherGuard } from "@/lib/guard";
-import { msg } from "@/msg-ja";
+import { msg } from "@/msg-com";
 
 const stringToBrightColor = (str: string) => {
     let hash = 0;
